LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scriptnonce-ignore-unsafeinline-expected.txt

ALERT: PASS (1/2)
ALERT: PASS (2/2)
CONSOLE MESSAGE: line 15: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'nonce-noncynonce' 'nonce-noncy+/=nonce' 'unsafe-inline'".

This tests that a valid nonce disables inline JavaScript, even if 'unsafe-inline' is present.
